+++ src/reducers/index.js
@@ -1,12 +1,10 @@
 import {combineReducers} from 'redux';
 import {routerReducer} from 'react-router-redux';
 
-import fuelSavings from './fuelSavingsReducer';
 import sideMenu from '../containers/SideMenu/SideMenu.reducer';
 import homePage from '../containers/Home/HomePage.reducer';
 
 const rootReducer = combineReducers({
-	fuelSavings,
 	sideMenu,
 	homePage,
 	routing: routerReducer